requirements.txt
torch==1.13.0+cu117
MinkowskiEngine==0.5.4
pytorch3d==0.7.7
nksr==1.0.3+pt113cu117
open3d
tensorboard
numpy
pyquaternion==0.9.9
python_pycg==0.5.2
PyYAML
scipy==1.8.1
tqdm
